+++ libIntegra/visualstudio/copyDependenciesToRelease.bat
@@ -14,14 +14,10 @@ copy "..\..\libIntegra_dependencies\lua-5.2.0\lua52.dll" "..\..\build\Release\se
 copy "..\data\CollectionSchema.xsd" "..\..\build\Release\server\"
 copy "..\data\id2guid.csv" "..\..\build\Release\server"
 
-rd /s /q ..\..\build\Release\modules
+if exist ..\..\build\Release\modules rd /s /q ..\..\build\Release\modules
 mkdir ..\..\build\Release\modules
 xcopy "..\..\modules" "..\..\build\Release\modules" /Y /Q
 
 copy "..\..\host\Pd\Integra_Host.pd" "..\..\build\Release\host\extra"
-
-rd /s /q ..\..\build\Release\gui\BlockLibrary
-mkdir ..\..\build\Release\gui\BlockLibrary
-xcopy "..\..\blocks" "..\..\build\Release\gui\BlockLibrary" /E /Y /Q
 
 CALL documentation_deployment\compileAllDocumentation.bat ..\..\build\Release\
